requirements.txt
pandas==2.2.3
numpy==1.26.4
yfinance==0.2.38
sqlalchemy==2.0.35
python-dotenv==1.0.1
streamlit==1.37.1
plotly==5.25.0